requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
SQLAlchemy==2.0.25
python-dotenv==1.0.0
APScheduler==3.10.4
